+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## XXX
+- Add optional fields to assay IImportRunOptions
+
 ## 1.1.4 - 2020-12-15
 - Add optional useAsync parameter to IImportDataOptions